refactor: move moviesUtils to utils
